LayoutTests/fast/scrolling/mac/wheel-event-in-overflow-with-shadow-dom-expected.txt


Test scroll over shadow content
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
Scrolling over this element should scroll the overflow scroller
